requirements.txt
python-telegram-bot[job-queue]==20.3
apscheduler==3.10.4
requests==2.31.0
python-dotenv==1.0.0
pytz==2023.3
feedparser==6.0.10
